requirements.txt
yt-dlp
python-telegram-bot==22.6
python-dotenv==1.2.1
ffmpeg-python==0.2.0
aiohttp==3.13.3
aiohttp-jinja2==1.6
jinja2==3.1.6
aiosqlite==0.21.0
